requirements.txt
opencv-python
pillow
pytesseract
numpy
torch
pdf2image
tqdm
